+++ client/lib/detect-window-boundary/index.ts
